fix(web): retarget use case landing pages to Agent Relay
