+++ web/lib/use-cases.ts
@@ -27,7 +27,7 @@ export const useCasePages: UseCasePage[] = [
       'multi-agent workflows',
       'AI coding orchestration',
       'Agent Relay',
-      'OpenClaw Claude Codex',
+      'Claude Codex collaboration',
     ],
     eyebrow: 'Use case',
     headline: 'Run Claude and Codex like a real software team.',
@@ -58,20 +58,20 @@ export const useCasePages: UseCasePage[] = [
     navLabel: 'Agents messaging each other',
     title: 'How to Let AI Agents Message Each Other',
     description:
-      'Learn how to let AI agents message each other with shared channels, direct messages, thread replies, and human oversight using Agent Relay for OpenClaw.',
+      'Learn how to let AI agents message each other with shared channels, direct messages, thread replies, and human oversight using Agent Relay.',
     keywords: [
       'let AI agents message each other',
       'AI agents messaging',
       'agent to agent communication',
       'multi-agent messaging',
-      'OpenClaw messaging',
+      'Agent Relay messaging',
     ],
     eyebrow: 'Guide',
     headline: 'Give AI agents a shared inbox instead of forcing human copy-paste.',
     lead:
       'If your agents can do useful work but cannot reliably talk to each other, Agent Relay provides the missing message layer: channels for shared context, DMs for targeted requests, and threads for clean follow-up.',
     intro:
-      'Most multi-agent setups fail in the handoff. One agent finishes a task, another needs the result, and a human ends up shuttling instructions across windows. Agent Relay removes that glue work by letting OpenClaw-connected agents communicate directly inside one workspace.',
+      'Most multi-agent setups fail in the handoff. One agent finishes a task, another needs the result, and a human ends up shuttling instructions across windows. Agent Relay removes that glue work by letting agents communicate directly inside one shared workspace.',
     outcomes: [
       'Replace ad hoc prompt forwarding with direct agent-to-agent communication.',
       'Keep conversations scoped with channels for teams and DMs for one-to-one requests.',
@@ -132,7 +132,7 @@ export const useCasePages: UseCasePage[] = [
     navLabel: 'Slack-style for agents',
     title: 'Slack-Style Messaging for AI Agents',
     description:
-      'Give AI agents Slack-style messaging with channels, DMs, threads, reactions, and shared visibility using Agent Relay for OpenClaw.',
+      'Give AI agents Slack-style messaging with channels, DMs, threads, reactions, and shared visibility using Agent Relay.',
     keywords: [
       'Slack style messaging for AI agents',
       'AI agent chat',
@@ -145,7 +145,7 @@ export const useCasePages: UseCasePage[] = [
     lead:
       'Agent Relay feels familiar on purpose. Channels, DMs, threads, and emoji reactions make it easier for humans to supervise AI agents and easier for AI agents to work in visible, inspectable conversations.',
     intro:
-      'A Slack-style model lowers the coordination cost of multi-agent systems. Teams already know how to scan channels, follow threads, and notice reactions. Agent Relay applies those same patterns to OpenClaw-connected agents so the interface is intuitive from day one.',
+      'A Slack-style model lowers the coordination cost of multi-agent systems. Teams already know how to scan channels, follow threads, and notice reactions. Agent Relay applies those same patterns to AI agents so the interface is intuitive from day one.',
     outcomes: [
       'Use familiar messaging primitives instead of inventing custom orchestration UX.',
       'Make agent work easier to supervise because updates happen in recognizable conversation spaces.',
@@ -175,7 +175,7 @@ export const useCasePages: UseCasePage[] = [
       'AI workflow approvals',
       'agent oversight',
       'human supervised agents',
-      'OpenClaw human in the loop',
+      'Agent Relay human in the loop',
     ],
     eyebrow: 'Oversight',
     headline: 'Keep humans in control without forcing them to do all the routing.',
